+++ src/econ/econ.py
@@ -1,7 +1,8 @@
 from src._road.road import OwnerID, PersonID, validate_roadnode
-from src._road.worldnox import EconNox, get_file_name
+from src._road.worldnox import get_file_name
 from src.agenda.party import partylink_shop
 from src.agenda.agenda import AgendaUnit, get_from_json as get_agenda_from_json
+from src.change.agendanox import EconNox
 from src.change.listen import get_speaker_agenda, listen_to_debtors_roll
 from src._instrument.file import set_dir, delete_dir, open_file, dir_files, save_file
 from dataclasses import dataclass
