+++ src/econ/examples/econ_env_kit.py
@@ -1,6 +1,6 @@
 from src._instrument.file import delete_dir, copy_dir
 from src._road.road import create_road_from_nodes, RoadUnit
-from src._road.worldnox import econnox_shop, EconNox
+from src.change.agendanox import econnox_shop, EconNox
 from os.path import exists as os_path_exists
 from pytest import fixture as pytest_fixture
