+++ src/world/system.py
@@ -1,4 +1,4 @@
-# from src._road.worldnox import econnox_shop
+# from src.change.agendanox import econnox_shop
 # from src._road.finance import default_planck_if_none
 # from src._road.road import (
 #     default_road_delimiter_if_none,
